Creates some of the folders required to organize outputs in the a certain provided path and prints time spent;

Creates some of the folders required to organize outputs in the a certain provided path and prints time spent;

CreateOutputFoldersALL(Year_List, OUTPUT_Path)

Arguments

Year_List

a list of integers representing the years to be calculated. Folders will be created based; If Windows based, MAXIMUM CHARACTER SIZE IS 118. Or go to the registry and change the maximum number of characters on a file path;

OUTPUT_Path

a string representing a path where you'd like the folders to be created;

Value

has no returns. Meant to be used just to create folders;