Takes a list of products, for example NDVI_2016, and cuts it to each individual ZOI, exporting the results to a certain path;

Takes a list of products, for example NDVI_2016, and cuts it to each individual ZOI, exporting the results to a certain path;

IndividualAlbufeiraProduct(
  ProductList_AllThreshold,
  AlbufeiraShapefileList,
  ExportPath,
  ID_list,
  StringName,
  StringOne,
  StringOneAlbufeira,
  StringAll
)

Arguments

ProductList_AllThreshold

a list of rasters representing each month of a particular product, example a list of each month of NDVI_2016;

AlbufeiraShapefileList

a list of shapefiles representing the individual ZOI of the original shapefile;

ExportPath

a string representing an export path;

ID_list

a list of IDs representing the amount of unique shapes inside the original shapefile, usually created asa global variable using CropShapefile_ExportToPath;

StringName

a string to name the files, usually XXX_ as the function will attach the name of the month in the end;

StringOne

a string to print as each product is finished;

StringOneAlbufeira

a string to print as each complete ZOI is finished;

StringAll

a string to print in the end;

Value

a list of lists, with the first list representing ZOI_1 and element 1 of the first list being the first month corresponding to the first ZOI;